+++ DrivingSchoolWebApp.Web/Controllers/CoursesController.cs
@@ -3,6 +3,7 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Security.Claims;
     using Data.Models.Enums;
     using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.Mvc;
@@ -48,9 +49,9 @@
                 }
                 else
                 {
-                    courses = this.courseService.GetCoursesByCategory<AllCoursesViewModel>( Enum.Parse<Category>(category)).ToList();
+                    courses = this.courseService.GetCoursesByCategory<AllCoursesViewModel>(Enum.Parse<Category>(category)).ToList();
                 }
-                
+
                 return this.View(courses);
             }
             catch (Exception e)
@@ -133,15 +134,22 @@
         // GET: Courses/Edit/5
         public ActionResult Edit(int id)
         {
-            var username = this.User.Identity.Name;
-            var schoolId = this.schoolService.GetSchoolByManagerName<SchoolViewModel>(username).Id;
-            this.ViewBag.Trainers = this.trainerService
-                .TrainersBySchoolId<AvailableTrainerViewModel>(
-                    schoolId);
-            this.ViewBag.Cars = this.carService.GetCarsBySchoolId<CarViewModel>(schoolId);
+            try
+            {
+                var username = this.User.Identity.Name;
+                var schoolId = this.schoolService.GetSchoolByManagerName<SchoolViewModel>(username).Id;
+                this.ViewBag.Trainers = this.trainerService
+                    .TrainersBySchoolId<AvailableTrainerViewModel>(
+                        schoolId);
+                this.ViewBag.Cars = this.carService.GetCarsBySchoolId<CarViewModel>(schoolId);
 
-            var model = this.courseService.GetCourseById<EditCourseInputModel>(id);
-            return this.View(model);
+                var model = this.courseService.GetCourseById<EditCourseInputModel>(id);
+                return this.View(model);
+            }
+            catch (Exception e)
+            {
+                return this.View("_Error", e.Message);
+            }
         }
 
         // POST: Courses/Edit/5
@@ -149,22 +157,42 @@
         [ValidateAntiForgeryToken]
         public ActionResult Edit(EditCourseInputModel model)
         {
-            if (!this.ModelState.IsValid)
+            try
             {
-                return this.View(model);
-            }
+                if (!this.ModelState.IsValid)
+                {
+                    return this.View(model);
+                }
 
-            model.Username = this.User.Identity.Name;
-            var course = this.courseService.Edit(model).GetAwaiter().GetResult();
-            return this.RedirectToAction("Details", "Courses", new { Area = "", id = course.Id });
+                if (!this.HasRights(model.Id))
+                {
+                    throw new OperationCanceledException("You do not have rights for this operation!");
+                }
+
+                model.Username = this.User.Identity.Name;
+                var course = this.courseService.Edit(model).GetAwaiter().GetResult();
+                return this.RedirectToAction("Details", "Courses", new { Area = "", id = course.Id });
+            }
+            catch (Exception e)
+            {
+                return this.View("_Error", e.Message);
+            }
         }
 
         // GET: Courses/Delete/5
         public ActionResult Delete(int id)
         {
-            var course = this.courseService.GetCourseById<DeleteCourseViewModel>(id);
+            try
+            {
+                var course = this.courseService.GetCourseById<DeleteCourseViewModel>(id);
 
-            return this.View(course);
+                return this.View(course);
+
+            }
+            catch (Exception e)
+            {
+                return this.View("_Error", e.Message);
+            }
         }
 
         // POST: Courses/Delete/5
@@ -174,6 +202,11 @@
         {
             try
             {
+                if (!this.HasRights(id))
+                {
+                    throw new OperationCanceledException("You do not have rights for this operation!");
+                }
+
                 this.courseService.Delete(id).GetAwaiter().GetResult();
 
                 return this.RedirectToAction(nameof(this.All));
@@ -182,6 +215,15 @@
             {
                 return this.View();
             }
+        }
+
+        private bool HasRights(int courseId)
+        {
+            var userId = this.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var course = this.courseService.GetCourseById<DetailsCourseViewModel>(courseId);
+
+            var result = (userId == course.SchoolManagerUserId) || this.User.IsInRole("Admin");
+            return result;
         }
     }
 }
